Added support for SoundCloud system-playlists.
SoundCloud has some playlists of the type "system-playlist". These are e.g. weekly mixes and other auto-generated lists. A playlist has to be public for its data to be fetched from the API, so in case you want to play e.g. your personal weekly mix via the music bot, make sure to change the list to public first.