Spotify fixes again...
Due to inconsistent Spotify API behaviour. it is sometimes possible for neither external_urls->spotify nor linked_from->external_urls->spotify to match the track's id.
As a workaround, grab the id and  create the track link manually.
linkedFrom is now included in all cases where a Track object is being created in case of differing links.

SongQueue will now check for the linkedFrom value too, in case either the bot or the player uses a different link.