SongQueue can now handle misbehaving players better.
Increased the amount of time the bot waits for the player when trying to play a track. Should help with issues on slower machines.
This was needed due to Spotify forcing oauth2 login now and players such as ncspot have longer startup times as a result.

SongQueue can now handle cases where a player suddenly stops playback either while a track is playing or paused. The bot then starts the song again and moves back to the same position.