SongQueue: fix exception when player returns no metadata
If the given media player returns no metadata when we run currentUrl(),
it would cause an exception because the code tried to parse it anyway.
Added ifOutputTextNotEmpty and ifErrorTextNotEmpty functions to the Output class.
These can be used to run code only if the output/error text is not empty.