services: added Apple Music support
The way it works is that when you add an apple music link to the queue,
it will fetch available links via SongLink and then choose a link according to our service priority:
1. Spotify, 2. Bandcamp, 3. YouTube and then will add that link to the queue.
This is because we can't play music from Apple Music directly.

Since SongLink uses Apple Music's search for its search functionality, added the same behaviour here too.